More refactoring and fixes
Add support for array concatenation using '+', and repeating an array using '*' with an int. Limit fuzzer memory usage so that the process isn't killed on some platforms (OOM killer). Don't set values to NULL in pawC_stkinc, as it is redundant. Get the CI to work on Windows, and modify the tests so that they don't use anything platform-specific (like /tmp).